refactor(budget): share the usage-delta helper across capabilities
The usage-delta pair - read four token counters off a shared `RunUsage`,
diff them to find what a nested agent call spent - lived twice, verbatim,
in compaction (`_counts`/`_spent`) and in the reminder capability added
by #806 (`_usage_counts`/`_usage_delta`), docstrings included. It is
pricing-sensitive, and two copies are two chances to fix one and miss the
other with a green suite - the exact drift the review of #806 flagged.

Hoist one copy into `budget` as `usage_counts` / `usage_delta`, next to
`record_ambient_usage` which both callers already use to book the spend,
and have compaction and system_reminders call it. No behaviour change;
the unit test moves to `test_spend.py` beside the helper.
