+++ backend/app/agents/capabilities/budget/__init__.py
@@ -13,6 +13,8 @@ from app.agents.capabilities.budget._capability import (
     metered_by,
     price_request,
     record_ambient_usage,
+    usage_counts,
+    usage_delta,
 )
 
 __all__ = [
@@ -28,4 +30,6 @@ __all__ = [
     "metered_by",
     "price_request",
     "record_ambient_usage",
+    "usage_counts",
+    "usage_delta",
 ]
